Changed scoring of Game 3
Instead of having 1 guess, the player can guess repeatedly which picture was only shown once. The pictures are then marked as (in)correct guesses with either a cross or a tick. The scoring for Game 3 (WhichPictureWasShownOnlyOnce) is in an ascending order - lowest number of guesses is the best - the PlayerScore class was modified accordingly.